treewide: general refactoring
