+++ eduu/plugins/gif.py
@@ -7,8 +7,7 @@ from pyrogram import Client, filters
 from pyrogram.types import Message
 
 from eduu.config import TENOR_API_KEY, prefix
-from eduu.utils import commands
-from eduu.utils.consts import http
+from eduu.utils import commands, http
 from eduu.utils.localization import use_chat_lang
 
 if not TENOR_API_KEY:
