+++ eduu/plugins/hastebin.py
@@ -5,8 +5,7 @@ from pyrogram import Client, filters
 from pyrogram.types import Message
 
 from eduu.config import prefix
-from eduu.utils import commands
-from eduu.utils.consts import http
+from eduu.utils import commands, http
 from eduu.utils.localization import use_chat_lang
