+++ eduu/utils/__init__.py
@@ -12,14 +12,15 @@ from .utils import (
     chat_exists,
     check_perms,
     commands,
-    deEmojify,
     del_restarted,
     get_emoji_regex,
     get_format_keys,
     get_restarted,
+    http,
     pretty_size,
     remove_escapes,
     require_admin,
+    run_async,
     set_restarted,
     shell_exec,
     split_quotes,
@@ -28,24 +29,25 @@ from .utils import (
 )
 
 __all__: List[str] = [
-    "commands",
-    "shell_exec",
-    "deEmojify",
-    "get_emoji_regex",
-    "button_parser",
-    "split_quotes",
-    "remove_escapes",
-    "time_extract",
-    "require_admin",
-    "check_perms",
-    "set_restarted",
-    "get_restarted",
-    "get_format_keys",
-    "del_restarted",
-    "chat_exists",
+    "EMOJI_PATTERN",
     "add_chat",
     "aiowrap",
+    "button_parser",
+    "chat_exists",
+    "check_perms",
+    "commands",
+    "del_restarted",
+    "get_emoji_regex",
+    "get_format_keys",
+    "get_restarted",
+    "http",
     "pretty_size",
-    "EMOJI_PATTERN",
+    "remove_escapes",
+    "require_admin",
+    "run_async",
+    "set_restarted",
+    "shell_exec",
+    "split_quotes",
     "sudofilter",
+    "time_extract",
 ]
