+++ eduu/utils/utils.py
@@ -2,6 +2,7 @@
 # Copyright (c) 2018-2022 Amano Team
 
 import asyncio
+import atexit
 import inspect
 import math
 import os.path
@@ -11,6 +12,7 @@ from functools import partial, wraps
 from string import Formatter
 from typing import Callable, List, Optional, Tuple, Union
 
+import httpx
 from pyrogram import Client, emoji, filters
 from pyrogram.types import CallbackQuery, InlineKeyboardButton, Message
 
@@ -29,7 +31,16 @@ BTN_URL_REGEX = re.compile(r"(\[([^\[]+?)\]\(buttonurl:(?:/{0,2})(.+?)(:same)?\)
 SMART_OPEN = "“"
 SMART_CLOSE = "”"
 START_CHAR = ("'", '"', SMART_OPEN)
-_EMOJI_REGEXP = None
+
+
+timeout = httpx.Timeout(40, pool=None)
+
+http = httpx.AsyncClient(http2=True, timeout=timeout)
+
+
+def run_async(func, *args, **kwargs):
+    loop = asyncio.get_event_loop()
+    loop.run_until_complete(func(*args, **kwargs))
 
 
 def pretty_size(size_bytes):
@@ -99,7 +110,6 @@ def set_restarted(chat_id: int, message_id: int):
 
 
 async def check_perms(
-    client: Client,
     message: Union[CallbackQuery, Message],
     permissions: Optional[Union[list, str]],
     complain_missing_perms: bool,
@@ -179,7 +189,7 @@ def require_admin(
             if msg.chat.type == "channel":
                 return await func(client, message, *args, *kwargs)
             has_perms = await check_perms(
-                client, message, permissions, complain_missing_perms, strings
+                message, permissions, complain_missing_perms, strings
             )
             if has_perms:
                 return await func(client, message, *args, *kwargs)
@@ -345,28 +355,20 @@ commands = BotCommands()
 
 
 def get_emoji_regex():
-    global _EMOJI_REGEXP
-    if not _EMOJI_REGEXP:
-        e_list = [
-            getattr(emoji, e).encode("unicode-escape").decode("ASCII")
-            for e in dir(emoji)
-            if not e.startswith("_")
-        ]
-        # to avoid re.error excluding char that start with '*'
-        e_sort = sorted([x for x in e_list if not x.startswith("*")], reverse=True)
-        # Sort emojis by length to make sure multi-character emojis are
-        # matched first
-        pattern_ = f"({'|'.join(e_sort)})"
-        _EMOJI_REGEXP = re.compile(pattern_)
-    return _EMOJI_REGEXP
+    e_list = [
+        getattr(emoji, e).encode("unicode-escape").decode("ASCII")
+        for e in dir(emoji)
+        if not e.startswith("_")
+    ]
+    # to avoid re.error excluding char that start with '*'
+    e_sort = sorted([x for x in e_list if not x.startswith("*")], reverse=True)
+    # Sort emojis by length to make sure multi-character emojis are
+    # matched first
+    pattern_ = f"({'|'.join(e_sort)})"
+    return re.compile(pattern_)
 
 
 EMOJI_PATTERN = get_emoji_regex()
-
-
-def deEmojify(text: str) -> str:
-    """Remove emojis and other non-safe characters from string."""
-    return EMOJI_PATTERN.sub("", text)
 
 
 # Thank github.com/usernein for shell_exec
@@ -384,3 +386,6 @@ async def shell_exec(code, treat=True):
 def get_format_keys(string: str) -> List[str]:
     """Return a list of formatting keys present in string."""
     return [i[1] for i in Formatter().parse(string) if i[1] is not None]
+
+
+atexit.register(run_async, http.aclose)
